requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.22.0
pandas>=1.4.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
pillow>=9.0.0
h5py>=3.7.0
astropy>=5.1.0
sunpy>=4.0.0
opencv-python>=4.5.0
pytorch-lightning>=1.8.0
transformers>=4.21.0
captum>=0.5.0
shap>=0.41.0
pyyaml>=6.0
pathlib>=1.0.1 